Extracted DeleteShouldIgnoreBulletsFeature
