+++ tests/features/DeleteShouldIgnoreBulletsFeature.test.ts
@@ -1,4 +1,4 @@
-import type ObsidianOutlinerPluginWithTests from "../test";
+import type ObsidianOutlinerPluginWithTests from "../../test";
 
 export default {
   "backspace should remove line if it's last empty line": (plugin) => {
@@ -20,6 +20,27 @@ export default {
 
     // assert
     plugin.assertCurrentState(["- qw|"]);
+  },
+  "backspace should remove list item if it's empty": (plugin) => {
+    // arrange
+    plugin.applyState(["- one", "- |"]);
+
+    // act
+    plugin.simulateKeydown("Backspace");
+
+    // assert
+    plugin.assertCurrentState(["- one|"]);
+  },
+  "cmd+backspace should remove content only": (plugin) => {
+    // arrange
+    plugin.applyState(["- one", "- two|"]);
+
+    // act
+    plugin.simulateKeydown("Cmd-Backspace");
+    plugin.simulateKeydown("Cmd-Backspace");
+
+    // assert
+    plugin.assertCurrentState(["- one", "- |"]);
   },
   "delete should remove next item if cursor is on the end": (plugin) => {
     // arrange
